+++ spec/integration/conversion_spec.rb
@@ -141,5 +141,47 @@ module Transpec
         end
       end
     end
+
+    describe 'one-liner expectation with have(n).errors_on(attr)' do
+      let(:source) do
+        <<-END
+          module RSpec
+            module Rails
+            end
+          end
+
+          module ActiveModel
+            module Validations
+              def errors_on(attribute, options = {})
+                valid_args = [options[:context]].compact
+                self.valid?(*valid_args)
+
+                [self.errors[attribute]].flatten.compact
+              end
+            end
+          end
+
+          class SomeModel
+            include ActiveModel::Validations
+
+            def valid?(*)
+              false
+            end
+
+            def errors
+              { name: [:foo, :bar] }
+            end
+          end
+
+          describe SomeModel do
+            it { should have(2).errors_on(:name) }
+          end
+        END
+      end
+
+      it "won't be converted" do
+        converted_source.should == source
+      end
+    end
   end
 end